Fix issue conjecture tokenizer
Apparently the tokenizer was disjoint from how the text was computed
for use in the dataset. This caused a disrepancy between the tokeniser
and the text used. The result was a vocbaulary consisisting mostly of unkown tokens.
This is of course a major issue, and it is unclear how long it has been like this.
Previously, the conjectures would run ok but not top performing. While with this
issue, the model learned nothing.